requirements.txt
anyio==3.6.1
asgiref==3.8.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Django
django-autoslug==1.9.8
django-extensions==3.2.1
django-model-utils==4.2.0
django-redis==5.2.0
django-seed==0.3.1
djangorestframework==3.15.2
drf-nested-routers==0.94.1
Faker==26.0.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
progressbar2==4.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
python-utils==3.3.3
pytz==2022.4
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
swapper==1.3.0
tomli==2.0.1
toposort==1.7
tqdm==4.64.1
Unidecode==1.3.6
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
wrapt==1.14.1
